Support GATK-compatible VCF output format.
There are 2 errors which can happen when running GATK Variant Annotator
with a SnpEff result file as input:
- in GATK2: "The version of SnpEff used to generate the SnpEff input
file (3.x) is not currently supported by the GATK, and was not run in
GATK compatibility mode. Supported versions are: [2.0.5], as well as
later versions when run with the option -o gatk": this is fixed just by
selecting the "GATK-compatible VCF" output type;
- in GATK1: "The version of SnpEff used to generate the SnpEff input
file (x.x) is not currently supported by the GATK. Supported versions
are: [2.0.5]": this is fixed by selecting the GATK-compatible VCF output
type and also the "Compatible with GATK 1.x" checkbox.