Check that multiselect is not None in output filters
Oddly, in Galaxy the value of an empty multiselect is None, not an empty
list. So, if the user selects no options, the evaluation of a filter like

    <filter>'gff' in outputs</filter>

will raise a TypeError exception and the dataset will not be filtered out.
See the filter_output() method in https://github.com/galaxyproject/galaxy/blob/dev/lib/galaxy/tools/actions/__init__.py