Basic syntax checks and code aeration
The wrapper is using the correct indentation type (4 spaces) and there is no indentation error. The value of the parameters in the "command" section are surrounded by single quotes. The "version_command", "citation" and "requirements" sections are present and their contents are Ok. The different sections are now separated by an empty line.