Filter_otus wrapper - Command line section update
The content of the command section has been updated to deal with the changes made in the inputs section. In addition, the syntax has been homogenized and updated to respect the guidelines. The "if $param != ''" syntax is now used instead of the "if $param" syntax cause the former better manages integers with value set to 0 during the command line building.