Filter_otus wrapper - Test files added to the test-data folder
A "filter_otus_from_otu_table" subfolder has been created in the qiime_core test-data folder. It contains 2 input files that comes from the qiime_test_data folder of the official Qiime repository as well as 3 output files generated with the generate_test_data script and that can be used to test the wrapper with planemo.