Get.oturep - Merge of the names and count parameters
In a similar way of what has been done in the pre.cluster tool, end users can now choose to use either a name file (mothur.names) or a count table (mothur.count_table) with a single parameter. The inputs section now have a nameOrCount param and all other sections have been updated accordingly. In the command section the appropriate command line option (name or count) is selected automatically base on the format of the selected file. In addition, a name/count file is now considered as mandatory in column mode (as stated in get.oturep(help)).