Get.oturep - Selection method switch
In the previous versions of the wrapper, the switch between the two available methods for representative sequences selection was considered as a secondary parameter. However, it should be consider as a main switch because most other options depends on it. This have been changed in this version and every section's content have been adapted.