Create.database - Merge of the repname and count parameters
In a similar way of what has been done in the pre.cluster tool, end users can now choose to use either a name file (mothur.names) or a count table (mothur.count_table) with a single parameter. The inputs section now have a nameOrCount param and in the command section the appropriate command line option (repname or count) is selected automatically base on the format of the selected file. In addition, the order of the parameters in every sections has been homogenized.