Make.contigs - Update of the outputs section
Since the name of the output files are based on the name of either the "ffastq" or "file" option of Mothur make.contigs then the value of the from_work_dir attributes must be slightly changed to match both cases. As a consequence, Galaxy is now looking for files matching the *fastq* pattern instead of the ffastq* pattern (The combo file is called combo_fastq.dat).