Make.contigs - New test for list:paired collections
A new <test> has been added in the tests section to be able to test the list:paired collection case with planemo. In addition, the syntax of the conditional params has been made more explicite and the descriptions of the tests have been slightly modified to become a bit clearer.