Assign_taxonomy - Activation/Deactivation of the assignment methods
After checking and testing it seems that the RDPtools conda package is not compatible with Qiime (Qiime determine if the RDP classifier is installed based on an environment variable that is not set) and there is no Rtax conda package at the moment. So, as a consequence the RDP and Rtax assignment methods are kept deactivated and the RDPtools requirements is commented.

On the opposite, the Blast, Mothur and SortMeRNA assignment methods are valid  choices and their requirements are fulfilled with the existing conda modules.