Assign_taxonomy - Multiple choices for the reference taxonomy file
The user can now choose the origin of the file provided to the --id_to_taxonomy_fp parameter. It can still comes from the history but it can also comes from an existing log file (qiime_taxonomy). Finally, the user can choose not to use this parameter since it is optional for certain assignment methods.
The command section has been updated accordingly.