Assign_taxonomy - Loc file support for the reference_seqs_fp
The user can now choose the origin of the file provided to the --reference_seqs_fp parameter. It can still come from the history but it can also come from an existing log file (qiime_rep_set). Since this parameter is used with several assignment methods, the definition of the conditional tag has been done in the macros.xml file. The macro is then called wherever it's needed in the assign_taxonomy wrapper.