Assign_taxonomy - Full update of the tests section
The syntax of the existing tests has been updated to be more explicit (with the conditional cases) and to reflect the modifications made in the previous commits of this branch.
In addition, new tests have been added for the Blast and Mothur assignment methods.
Moreover, the comparisons of the expected and real output files are now made based on md5 checksums or assert statements instead of full file comparisons.
Finally, the SortMeRNA test is kept commented by default since it's execution is a bit too long even on small test files (However, it has been executed with planemo and pass without any error).