Assign_taxonomy - Generate_test_data update
The generate_test_data.sh shell script has been modified to add the command lines of the new and updated tests of the Assign_taxonomy tool.
Since all the tests are now based on md5 checksum comparisons or assert statements the content of the generated folders is now just displayed as well as the md5 checsum of the relevant files.