Assign_taxonomy - Smaller input test file for the SortMeRNA test
The SortMeRNA test takes more than 20 minutes with the current input file and set of parameters. The issue is that this execution time exceed the timeout set for the project and makes Travis builds fail. To avoid that, I've reduced the size of the input file used during this specific test.