Assign_taxonomy - Deactivation of the --sortmerna_db parameter
This parameter is currently broken since it asked for an unindexed fasta database to the user while the Qiime assign_taxonomy.py script is waiting for the name of a database index. The corresponding param tag has been commented out (Note that the value of the argument attribute has been slightly modified to avoid a syntax error due to the "--" string it contains).