Add GNU parallel citation
I saw the "--will-cite" flag passed to scripts and wondered what it was. 

https://www.gnu.org/software/parallel/man.html
```
If you use --will-cite in scripts to be run by others you are making it harder for others to see the citation notice. The development of GNU parallel is indirectly financed through citations, so if your users do not know they should cite then you are making it harder to finance development. However, if you pay 10000 EUR, you should feel free to use --will-cite in scripts.
```
I assume we didn't pay 10K EUR, so it seems reasonable to add this citation (also to any other tools that use parallel).