Use element identifier to label samples and use pheatmap
We use the element_identifier for labels, so that all plots and the normalized
counts table are properly labeled. This fixes https://github.com/galaxyproject/tools-iuc/issues/1484.

Using pheatmap gives a nicer heatmap that scales better for few/many samples (and is used in the [DESeq2 vignette](https://bioconductor.org/packages/release/bioc/vignettes/DESeq2/inst/doc/DESeq2.html#heatmap-of-the-sample-to-sample-distances))