Fix typo in Gene Body Coverage tool name?
Can tool names be changed? Can this tool name be changed to `Gene Body Coverage (BAM)` instead of the current name that has a typo (`Gene Body Converage (BAM)`) ? As `Gene Body Converage (BAM)` is what is used as the collection name in the history.

Same typo is in the related `Gene Body Converage (Bigwig)` tool.