Avoid blacklist/whitelist in user-facing tool text
The terms 'whitelist' and 'blacklist' for inclusion or exclusion
have undesirable associations, making it better to use more neutral
language.

See also "Be careful in the words that we choose" in the main
Galaxy Project Code of Conduct,
https://github.com/galaxyproject/galaxy/blob/dev/CODE_OF_CONDUCT.md

We can't change the parameter names without breaking backward
compatibility, so this is a cosmetic change only. The command
line arguments -B and -W remain, but this reflects the underlying
tool itself.