Fix option ${save_raw} in <command/>
If boolean ${opt} is set in the recommended way, i.e. "--opt" when
true and "" when false, then add single quotes around, i.e. '${opt}'
will break the executed command, as argparse will treat '' as an
empty positional argument which it doesn't recognise. This fix simply
removes the double quote as a temporary measure. Better ways need to
be sought.