Use new group_tag parameter to allow selecting collection subsets
The idea here is that the user tags elements of a collection using tags that
start with `group:`. All of these tags can then be selected in the tool
form when choosing the factor levels.
This makes it very easy to select dataset combinations of interest
without having to re-shape collection contents for every contrast.

Technically the input does not need to be a collection, a
multiple="true" data input also works, but is currently not testable.
We might change this in the future.

The previous way of doing the analysis is still the default way of
using deseq2, as we currently don't auto-propagate group tags.
So in order to use this new feature users either have to manually tag collection
elements or use the `Tag elements from file from contents of a file`
to set the group tags. Hopefully we will enable auto-propagation
in 19.01 and enhance the apply rules tool to manipulate group tags.