Enable testing of gemini_load
The gemini_load wrapper depends on the gemini annotation files
being available. Due to the size of GEMINI's native annotation
files the wrapper could not be tested so far, but this changeset
adds downsampled annotation files and utility scripts to generate
them for any given version of gemini, and introduces a working
test for gemini_load.