Fix the Only annotate ID field option of SnpSift Annotate
From snpSift v4.1.h onwards, the -noInfo option is used instead
of the -id option of the annotate command is used to indicate that
the INFO field of the input VCF should be left untouched.

Unfortunately, the snpSift documentation was never updated to
reflect that change and the Galaxy wrapper was never changed.