bcftools suite enhancements (#2440)
* Enable cached genomes throughout bcftools suite

The tools:
- bcftools_consensus
- bcftools_convert_to_vcf
- bcftools_csq
- bcftools_norm
do now accept a cached indexed fasta reference as an alternative
to a fasta dataset. Corresponding tool tests have been added.

In addition, the changeset improves the input mask for specifying
regions or targets restrictions, and improves the interface of
bcftools_norm, in particular.

* Add cached genome support to bcftools_stats

* Fix and add test for frameshifts plugin

* Fix and add tests for color-chrs plugin

Among other things, removes the previous failed attempt to support
the --colors option of color-chrs.pl since implementing it in a
useful way requires a lot more effort.

* Fix and add tests for bcftools cnv

* Turn matplotlib requirements into macro

Also eliminates unused stdio macro.

* Bump mpileup version once more to enable rebase on master branch

* Make cached genome symlinks relative

* Remove ``exclude:`` lines from ``.shed.yml``

They do not prevent the creation of repositories for the listed
tools, just create repos with no tools inside.

* test against 19.05

* Use lowercase parameter names and fix quoting

* Update suite version info

Removes version info from the suite name and updates the version
in the description.

* Provide a more useful way to specify genomic intervals

With this change genomic intervals for the --regions and --targets
options can be specified without a start and/or stop position and
any missing position will be interpreted as "up to the
corresponding end of the chromosome".