Extend the whitelists of safe expression elements
The built in functions min, max and sum, the comma, and the .T
methods of pd.DataFrames and pd.Series have been added to the lists
of safe expression elements. This is necessary to make the examples
advertized in the tool help work.
In addition, runs of opening or closing parentheses are now parsed
as individual tokens to make parsing of nested function calls
robust.