Disallow use of expression attributes
A previous extension of the tool allowed the use of attributes of
expressions (like in (table.T - table.mean(1)).T) guessing the
expression's return type. This weakens parsing stringency
considerable and so gets disallowed by this changeset again.
Instead the general workaround - running the tool twice and
generating an intermediate table - is emphasized more in the tool
help.