travis: test for mixed tool data_manager PR
if there is more than 1 changed repo (eg 1 tool and 1 data manager)
then planemo is called per repo, which fails for data manager because
the path to the xml is needed.

hence if the repo is in the data_managers directory the path to the
xml file is determined with an additional call to `planemo
ci_find_tools` and then used in the planemo test call.