Make subsampling random and fix when subsample >= sample
awk's rand() function needs to be seeded to return different series of
random numbers. This changeset switches to using Python's
random.randrange() to generate truely random values for the seed part of
the samtools view -s option.
It also fixes a bug that caused the tool to fail for target numbers of reads
higher than the number found in the input.