Restructure samtools view
The main purpose of this major change to the tool interface is to
present only meaningful combinations of options to the user.

Examples of such changes include:

- all filter and subsampling options are no longer shown if the user
  asks for just the input header
- options to produce a dropped reads output are not offered in read
  count mode
- options to modify reads (-x and -B) are not offered in read count
  mode
- the choice to include the header in the output is only given when
  the output format is SAM

In addition, the restructured tool interface gives the possibility
to report dropped reads *instead* of retained ones in the single
output dataset, and moves the -x command line option out of the
filter and into the output options section (-x doesn't act as a
filter).