Fix lofreq indel call options and help
This changeset fixes the broken --only-indels option of lofreq call.
It also improves the tool and param help to make it clearer that indels
cannot be called with lofreq without adding indel qualities first with
lofreq indelqual (the mention of GATK as an alternative is outdated
since new versions of GATK don't generate indel quality tags anymore,
and got removed).