Update query_tabular for empty datasets
Allow query_tabular to create empty tables if possible.   This can be useful in workflows.

If there are no data rows in an input tabular dataset, query_tabular will fail unless:

a) There is a row header line in the dataset and Use first line as column names is selected
b) Specify Column Names is used to provide column names

in which case an empty table will be created with those columns.