Allow MultiQC runs with certain quasi-empty inputs
The specific cases fixed here are:
- bowtie2 mapping stats generated from an empty fastq
- picard MarkDuplicates stats generated from empty bams

In both cases, the MultiQC wrapper failed because it was looking for
missing text patterns, while MultiQC itself is perfectly
capable of dealing with these inputs (and will simply produce empty
output).
The issue is fixed here by scanning for parts of text that are constant
under all conditions.