IQtree: --mset parameter requires lowercase (#3927)
Default values are currently unable to be used on the --mset parameter. The ones that appear in the help text are "(raxml, phyml or mrbayes)". Lower case is required, and they are being sanitized away with the current version. Not bumping version, since this is a bug in the wrapper.