add -nameOnly flag
The "-nameOnly" flag differs from the "-name" flag by adding only the content of the 4th column in the bed file to the output fasta header. The "-name" flag adds both the name and the coordinates. Some tools require very specific requirements for the fasta header and there -nameOnly appears to be very convenient.