bedtools GetFastaBed: add -nameOnly flag (#4323)
* add -nameOnly flag

The "-nameOnly" flag differs from the "-name" flag by adding only the content of the 4th column in the bed file to the output fasta header. The "-name" flag adds both the name and the coordinates. Some tools require very specific requirements for the fasta header and there -nameOnly appears to be very convenient.

* added a "=True" test

* added more tests and changed the tool name

Now 4 tests are executed:
- fasta output, nameOnly FALSE
- tab output, nameOnly FALSE
- fasta output, nameOnly TRUE
- tab output, nameOnly TRUE

* added test data for getfasta