Make guppyplex input handling more flexible
This allows the user to specify which sequence input files guppyplex
should collapse into a single output, and which ones it should keep
separate. Also makes the tool work with fastq and fastq.gz data alike
and exposes the --quality threshold parameter.

The changeset also contains a fix for the artic minion pipeline, which
should make it handle input data with spaces in its name.