Update to latest version and fix error in some input names (#4397)
The tool does not end correctly (no error but empty outputs) if the names of the datasets could be interpreted as numbers (e.g. `001`). It is a bug in the upstream tool, reported and probably fixed.
Until a new release is available, quick fix in the wrapper has been added: adding `.fasta` to the input names during the symlink step.