Handle latest UShER assignment cache
The issue here is that pangolin's current logic uncondtionally installs
the assignment cache data into the conda environment.
This is not only bad in the context of Galaxy, but also the tool then
fails to update it in there, i.e. once installed the version is frozen
and trying to use the latest cache will result in an error.