concoct: fix test
fails CI with

```
Output process_log:  different than expected
Expected file size of 786+-0 found 787
```

I guess this is because the log contains the job working dir, eg
`/tmp/tmpnp9yg8__/job_working_directory/000/3/working/outdir`.
if the job is executed later the jobid may be larger than 3, ie
the log gets longer.

Assuming that the part that is `3` in the example can not
get longer than 5 digits I allow a delta of 4