add qiime qzv visualisation extractor
the qiime2 display application (view.qiime2.org) does not work for non-public Galaxy
instances. This is a workaround.

qiime2 viz artifacts are just zip files which also contain the html + data + files
(which is shown at the view website). So this tool is equivalent

If I find the time I will explore the possibility to do the same with a
generic / viz plugin .. at least for so long (and maybe for workflows)
the tool will be useful.