cutadapt name collusion
Added a check against name collusion between two input files. This becomes a problem if the tool is part of a workflow that provides a list of list of read pairs. The current version does not generate an error or warning, but rather generates inaccurate results by overwriting input1 with input2.