Add map_param_value expression tool
This will be very helpful for the 5th bullet point in https://github.com/galaxyproject/iwc/pull/126:
> Also, if there is a way in galaxy to do a:
if $strandness is 'unstranded' then parameter value is blabla elif $strandess is 'forward' then parameter value is bliblibli else parameter value is blobloblo' This would be better than letting the user choose between 3 awk commands and this would help for the stranded coverage.

I think we will eventually want to allow attaching arbitrary expressions
to steps, which have access to a particular steps' jobs object,  closer
to what CWL allows, but in the meantime this can be used to have
user-friendly text values as workflow inputs.