Use gemini's python in its data manager
This specifies 18.09 as the profile of the wrapper to prevent Galaxy's
Python from being used. It also changes the DM script to run the actual
download last (so that script errors will surface immediately instead of
after a download of several GBs), and adds tests and a test mode that
does not actually download anything.