fix arriba test (#4984)
the filter of output aligned_bam was broken (the used variables just do
not exist). this led to an error because Galaxy tried to set metadata
for and empty or non-existant file.

the output was removed because its not needed (its essentially the same
as the input). it probably stems from an earlier version that included
the alignment step.

also the alignment seems to be used if

- $output_fusion_bams and
- str($visualization.do_viz)=="yes"

so it needs to be generated in both