cutadapt: only allow fastqsanger
if we would want to consider fastqillumina/solexa then we would need
to set `--quality-base 33`. while this would be easy, we would also
need to set the output type equal to the input type (which seems
not worth the effort) .. alternatively we could call some converter
on the output to make it fastqsanger.