Update fastp.xml
Setting "report_json" to "True"

fastp should really replace fastqc for good. To do this we need to explain users that json report can be readily visualized with multiqc, but for that it needs to be enabled. 

Enabling it by default will eliminate the needs to tell users to rerun their fastp runs to generate the reports.